backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "AI slop",email = "farhanauliandaa@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi[standard] (>=0.133.1,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sse-starlette (>=3.2.0,<4.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "strands-agents (>=1.28.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "poethepoet (>=0.42.1,<0.43.0)",
    "faker (>=40.5.1,<41.0.0)"
]

[tool.poe.tasks]
dev = "fastapi dev main.py"
